analyze_db/requirements.txt
rich>=13.7.0
python-dotenv
pydantic_ai
fastapi>=0.109.0
uvicorn>=0.27.0
